requirements.txt
backports.shutil-get-terminal-size==1.0.0
colorama==0.3.9
decorator==4.3.0
Django==1.11.15
djangorestframework==3.8.2
enum34==1.1.6
ipdb==0.11
ipython==5.8.0
ipython-genutils==0.2.0
pathlib2==2.3.2
pickleshare==0.7.4
prompt-toolkit==1.0.15
Pygments==2.2.0
pytz==2018.5
scandir==1.9.0
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
wcwidth==0.1.7
win-unicode-console==0.5
